fix: address PR #40 review issues in network health indicators
- Fix pruneStaleNodes timing bug: revert to currentTimeMillis() instead
  of lastPollTimeMs which created stale-time-at-poll-start semantics
- Make lastPollTimeMs thread-safe with atomicfu atomic() since it is
  written by the poll loop and read by the listen loop
- Fix MascotAlert touch blocking: wrap Box in null check so fillMaxSize
  does not intercept touches when no alert is visible
- Deduplicate formatUptime() and DetailRow() from NodeCard and
  NetworkHealthTopBar into shared NetworkUiUtil.kt
- Replace wildcard imports with explicit imports in MascotAlert,
  NetworkHealthTopBar, and NodeCard
- Replace fully-qualified references with proper imports in
  NodeHealthIndicator (Size, Stroke)
