+++ shared/src/commonMain/kotlin/com/chromadmx/ui/screen/network/NodeHealthIndicator.kt
@@ -5,6 +5,8 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.dp
@@ -77,7 +79,7 @@ fun PixelHeart(
             drawRect(
                 color = color,
                 topLeft = Offset(x * pixelSize, y * pixelSize),
-                size = androidx.compose.ui.geometry.Size(pixelSize, pixelSize)
+                size = Size(pixelSize, pixelSize)
             )
         }
 
@@ -85,8 +87,8 @@ fun PixelHeart(
             drawRect(
                 color = color.copy(alpha = 0.5f),
                 topLeft = Offset(x * pixelSize, y * pixelSize),
-                size = androidx.compose.ui.geometry.Size(pixelSize, pixelSize),
-                style = androidx.compose.ui.graphics.drawscope.Stroke(width = 1f)
+                size = Size(pixelSize, pixelSize),
+                style = Stroke(width = 1f)
             )
         }
